Docs and extract replace to separate file
